requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cuda-python==12.6.0
cudf-cu12==24.12.0
cugraph-cu12==24.12.0
cuml-cu12==24.12.0
cupy-cuda12x==13.3.0
cuvs-cu12==24.12.0
cycler==0.12.1
dask==2024.11.2
dask-cuda==24.12.0
dask-cudf-cu12==24.12.0
dask-expr==1.1.19
debugpy==1.8.11
decorator==5.1.1
distributed==2024.11.2
distributed-ucxx-cu12==0.41.0
executing==2.1.0
fastrlock==0.8.3
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.68.1
h5py==3.12.1
hdbscan==0.8.40
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
libcudf-cu12==24.12.0
libkvikio-cu12==24.12.1
libucx-cu12==1.17.0.post1
libucxx-cu12==0.41.0
llvmlite==0.43.0
locket==1.0.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
narwhals==1.19.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numba-cuda==0.0.17.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu11==2.21.5
nvidia-nvcomp-cu12==4.1.0.6
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvtx==0.2.10
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prince==0.14.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
Pygments==2.18.0
pylibcudf-cu12==24.12.0
pylibcugraph-cu12==24.12.0
pylibraft-cu12==24.12.0
pynvjitlink-cu12==0.4.0
pynvml==11.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
raft-dask-cu12==24.12.0
rapids-dask-dependency==24.12.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rmm-cu12==24.12.1
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==70.0.0
six==1.17.0
sklearn-compat==0.1.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
tblib==3.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
treelite==4.3.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ucx-py-cu12==0.41.0
ucxx-cu12==0.41.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
zict==3.0.0
